requirements.txt
Flask==2.1.1
SQLAlchemy==1.4.35
WTForms==3.0.1
Werkzeug==2.1.1
Pillow==9.1.0
requests==2.27.1
